Fully replaced output with OutputHelper
